Review fixes
Removed car storage from factory class, renamed the factory class
