+++ homework2/CarFactory/CarFactory/Factories/ICarsFactory.cs
@@ -1,0 +1,9 @@
+﻿using CarFactory.Factories.Cars;
+
+namespace CarFactory.Factories
+{
+    public interface ICarsFactory
+    {
+        public Car CreateNewCar();
+    }
+}
