+++ homework2/CarFactory/CarFactory/Program.cs
@@ -1,12 +1,13 @@
 ﻿using CarFactory.Factories;
+using CarFactory.Factories.Cars;
 namespace CarFactory
 {
     public partial class Program
     {
         public static void Main()
         {
-            var factory = new Factory();
-            factory.CreateNewCar();
+            var carsFactory = new CarsFactory();
+            List<Car> carList = [carsFactory.CreateNewCar()];
             string strIndex;
             int index;
             bool exit = false;
@@ -27,16 +28,23 @@ namespace CarFactory
                 }
                 switch (index)
                 {
-                    case 1: factory.CreateNewCar(); break;
-                    case 2: factory.ListCars(); break;
+                    case 1: carList.Add(carsFactory.CreateNewCar()); break;
+                    case 2:
+                        foreach (Car car in carList)
+                        {
+                            Console.WriteLine($"{carList.IndexOf(car) + 1}. {car.Name}, {car.Model.Name}");
+                        }
+                        Console.ReadLine();
+                        break;
                     case 3:
                         Console.WriteLine("Enter index of the car:");
                         success = int.TryParse(Console.ReadLine(), out index);
-                        while (!success && (index <= 0 || index > factory.CarList.Count))
+                        while (!success && (index <= 0 || index > carList.Count))
                         {
                             success = int.TryParse(Console.ReadLine(), out index);
                         }
-                        factory.ListCarInfo(index);
+                        Console.WriteLine(carList[index - 1].ToString());
+                        Console.ReadLine();
                         break;
                     case 4:
                         exit = true;
